move button Export History to history pane
