+++ gui/wxpython/gui_core/goutput.py
@@ -29,7 +29,6 @@ from grass.pydispatch.signal import Signal
 from grass.grassdb.history import (
     read_history,
     create_history_file,
-    copy_history,
     get_current_mapset_gui_history_path,
 )
 
@@ -164,19 +163,10 @@ class GConsoleWindow(wx.SplitterWindow):
         self.btnOutputSave.SetToolTip(_("Save output to a file"))
         self.btnCmdAbort = Button(parent=self.panelProgress, id=wx.ID_STOP)
         self.btnCmdAbort.SetToolTip(_("Abort running command"))
-        self.btnCmdExportHistory = Button(parent=self.panelPrompt, id=wx.ID_ANY)
-        self.btnCmdExportHistory.SetLabel(_("&Export history"))
-        self.btnCmdExportHistory.SetToolTip(
-            _("Export history of executed commands to a file")
-        )
-
-        if not self._gcstyle & GC_PROMPT:
-            self.btnCmdExportHistory.Hide()
 
         self.btnClear.Bind(wx.EVT_BUTTON, self.OnClear)
         self.btnOutputSave.Bind(wx.EVT_BUTTON, self.OnOutputSave)
         self.btnCmdAbort.Bind(wx.EVT_BUTTON, self._gconsole.OnCmdAbort)
-        self.btnCmdExportHistory.Bind(wx.EVT_BUTTON, self.OnCmdExportHistory)
 
         self._layout()
 
@@ -206,12 +196,6 @@ class GConsoleWindow(wx.SplitterWindow):
             btnSizer = wx.BoxSizer(wx.HORIZONTAL)
             btnSizer.Add(
                 self.btnOutputSave,
-                proportion=0,
-                flag=wx.EXPAND | wx.LEFT | wx.RIGHT,
-                border=5,
-            )
-            btnSizer.Add(
-                self.btnCmdExportHistory,
                 proportion=0,
                 flag=wx.EXPAND | wx.LEFT | wx.RIGHT,
                 border=5,
@@ -447,33 +431,6 @@ class GConsoleWindow(wx.SplitterWindow):
     def OnCmdProgress(self, event):
         """Update progress message info"""
         self.progressbar.SetValue(event.value)
-        event.Skip()
-
-    def OnCmdExportHistory(self, event):
-        """Export the history of executed commands stored
-        in a .wxgui_history file to a selected file."""
-        dlg = wx.FileDialog(
-            self,
-            message=_("Save file as..."),
-            defaultFile="grass_cmd_log.txt",
-            wildcard=_("{txt} (*.txt)|*.txt|{files} (*)|*").format(
-                txt=_("Text files"), files=_("Files")
-            ),
-            style=wx.FD_SAVE | wx.FD_OVERWRITE_PROMPT,
-        )
-
-        if dlg.ShowModal() == wx.ID_OK:
-            path = dlg.GetPath()
-            history_path = get_current_mapset_gui_history_path()
-            try:
-                copy_history(path, history_path)
-                self.showNotification.emit(
-                    message=_("Command history saved to '{}'".format(path))
-                )
-            except OSError as e:
-                GError(str(e))
-
-        dlg.Destroy()
         event.Skip()
 
     def OnCmdRun(self, event):
