+++ gui/wxpython/history/browser.py
@@ -19,10 +19,16 @@ for details.
 
 import wx
 
-from gui_core.wrap import SearchCtrl
+from gui_core.wrap import SearchCtrl, Button
 from history.tree import HistoryBrowserTree
 
 from grass.pydispatch.signal import Signal
+from grass.grassdb.history import (
+    copy_history,
+    get_current_mapset_gui_history_path,
+)
+
+from core.gcmd import GError
 
 
 class HistoryBrowser(wx.Panel):
@@ -57,6 +63,14 @@ class HistoryBrowser(wx.Panel):
         self.search.Bind(wx.EVT_TEXT, lambda evt: self.tree.Filter(evt.GetString()))
         self.search.Bind(wx.EVT_SEARCHCTRL_CANCEL_BTN, lambda evt: self.tree.Filter(""))
 
+        # buttons
+        self.btnCmdExportHistory = Button(parent=self, id=wx.ID_ANY)
+        self.btnCmdExportHistory.SetLabel(_("&Export history"))
+        self.btnCmdExportHistory.SetToolTip(
+            _("Export history of executed commands to a file")
+        )
+        self.btnCmdExportHistory.Bind(wx.EVT_BUTTON, self.OnCmdExportHistory)
+
         # tree with layers
         self.tree = HistoryBrowserTree(self, giface=giface)
         self.tree.SetToolTip(_("Double-click to run selected tool"))
@@ -68,12 +82,11 @@ class HistoryBrowser(wx.Panel):
     def _layout(self):
         """Dialog layout"""
         sizer = wx.BoxSizer(wx.VERTICAL)
-        sizer.Add(
-            self.search,
-            proportion=0,
-            flag=wx.ALL | wx.EXPAND,
-            border=5,
-        )
+        searchSizer = wx.BoxSizer(wx.HORIZONTAL)
+        searchSizer.Add(self.search, proportion=1, flag=wx.EXPAND | wx.RIGHT, border=5)
+        searchSizer.Add(self.btnCmdExportHistory, proportion=0, flag=wx.EXPAND)
+
+        sizer.Add(searchSizer, proportion=0, flag=wx.EXPAND | wx.ALL, border=5)
         sizer.Add(
             self.tree, proportion=1, flag=wx.EXPAND | wx.LEFT | wx.RIGHT, border=5
         )
@@ -81,3 +94,30 @@ class HistoryBrowser(wx.Panel):
         self.SetSizerAndFit(sizer)
         self.SetAutoLayout(True)
         self.Layout()
+
+    def OnCmdExportHistory(self, event):
+        """Export the history of executed commands stored
+        in a .wxgui_history file to a selected file."""
+        dlg = wx.FileDialog(
+            self,
+            message=_("Save file as..."),
+            defaultFile="grass_cmd_log.txt",
+            wildcard=_("{txt} (*.txt)|*.txt|{files} (*)|*").format(
+                txt=_("Text files"), files=_("Files")
+            ),
+            style=wx.FD_SAVE | wx.FD_OVERWRITE_PROMPT,
+        )
+
+        if dlg.ShowModal() == wx.ID_OK:
+            path = dlg.GetPath()
+            history_path = get_current_mapset_gui_history_path()
+            try:
+                copy_history(path, history_path)
+                self.showNotification.emit(
+                    message=_("Command history saved to '{}'".format(path))
+                )
+            except OSError as e:
+                GError(str(e))
+
+        dlg.Destroy()
+        event.Skip()
